+++ vi/serversideaction.py
@@ -1,11 +1,11 @@
-# -*- coding: utf-8 -*-
-
 import logging
 
 from flare.button import Button
 from flare.html5 import window
-from flare.network import NetworkService
+from flare.network import NetworkService, NiceErrorAndThen
 from flare.safeeval import SafeEval
+from flare.i18n import translate
+from flare.popup import Confirm
 
 from vi.config import conf
 
@@ -17,6 +17,7 @@ class ServerSideActionWdg(Button):
 		self.handler = handler
 		self.actionName = actionName
 		self.actionData = actionData
+		self.multiSelection = self.actionData.get("allowMultiSelection", False)
 		self["class"] = "bar-item btn btn--small %s" % actionData["icon"]
 		self["disabled"] = True
 		self.isDisabled = True
@@ -51,11 +52,11 @@ class ServerSideActionWdg(Button):
 		super(ServerSideActionWdg, self).onDetach()
 
 	def onSelectionChanged(self, table, selection, *args, **kwargs):
-		if self.actionData["enabled"] == "True":
+		if self.actionData.get("enabled") == "True":
 			self.switchDisabledState(False)
 			return
 
-		if self.actionData["allowMultiSelection"] and len(selection) > 0 or len(selection) == 1:
+		if self.multiSelection and len(selection) > 0 or len(selection) == 1:
 			if not self.pendingFetches:
 				if self.selectionCheckerAst:
 					valid = True
@@ -76,35 +77,64 @@ class ServerSideActionWdg(Button):
 			self.switchDisabledState(True)
 
 	def onClick(self, sender=None):
+		if self.actionData.get("confirm",None) and self.actionData["confirm"]:
+			Confirm(self.actionData["confirm"],
+					title=translate(self.actionData['name']),
+					yesCallback=self.apply)
+		else:
+			self.apply()
+
+
+	def apply(self,sender=None):
 		selection = self.parent().parent().getCurrentSelection()
-		if self.actionData["allowMultiSelection"] and len(selection) > 0 or len(selection) == 1:
-			# if (len(selection) == 1 and not self.actionData["allowMultiSelection"]) or len(selection) > 0:
-			wasIdle = not self.pendingFetches
-			for item in selection:
+		if self.actionData["action"] == "view":
+			url_parts = self.actionData['url'].split("/")
+			targetInfo = conf["modules"][url_parts[0]]
+
+			if "params" in self.actionData:
+				if selection:
+					targetInfo.update({k: v.replace("{{key}}", selection[0]["key"]) for k, v in self.actionData["params"].items()})
+				else:
+					targetInfo.update(self.actionData["params"])
+			conf["mainWindow"].openView(
+				translate("{{module}} - {{name}}", module=targetInfo["name"], name=self.actionData['name']),
+				targetInfo.get("icon") or "icon-edit",
+				targetInfo["moduleName"] + targetInfo["handler"],
+				targetInfo["moduleName"],
+				None,  # is not used...
+				data=targetInfo
+				#data=utils.mergeDict(self.adminInfo, {"context": context})
+			)
+
+		else:
+			if self.multiSelection and len(selection) > 0 or len(selection) == 1:
+				# if (len(selection) == 1 and not self.actionData["allowMultiSelection"]) or len(selection) > 0:
+				wasIdle = not self.pendingFetches
+				for item in selection:
+					if self.actionData["action"] == "open":
+						url = self.actionData["url"].replace("{{key}}", item["key"])
+						window.open(url, "_blank")
+					elif self.actionData["action"] == "fetch":
+						url = self.actionData["url"].replace("{{key}}", item["key"])
+						self.pendingFetches.append(url)
+					else:
+						raise NotImplementedError()
+				if wasIdle and self.pendingFetches:
+					self.addClass("is-loading")
+					self.fetchNext()
+			elif self.actionData["enabled"] == "True":
+				wasIdle = not self.pendingFetches
 				if self.actionData["action"] == "open":
-					url = self.actionData["url"].replace("{{key}}", item["key"])
+					url = self.actionData["url"]
 					window.open(url, "_blank")
 				elif self.actionData["action"] == "fetch":
-					url = self.actionData["url"].replace("{{key}}", item["key"])
+					url = self.actionData["url"]
 					self.pendingFetches.append(url)
 				else:
 					raise NotImplementedError()
-			if wasIdle and self.pendingFetches:
-				self.addClass("is-loading")
-				self.fetchNext()
-		elif self.actionData["enabled"] == "True":
-			wasIdle = not self.pendingFetches
-			if self.actionData["action"] == "open":
-				url = self.actionData["url"]
-				window.open(url, "_blank")
-			elif self.actionData["action"] == "fetch":
-				url = self.actionData["url"]
-				self.pendingFetches.append(url)
-			else:
-				raise NotImplementedError()
-			if wasIdle and self.pendingFetches:
-				self.addClass("is-loading")
-				self.fetchNext()
+				if wasIdle and self.pendingFetches:
+					self.addClass("is-loading")
+					self.fetchNext()
 
 	def fetchNext(self):
 		if not self.pendingFetches:
@@ -113,21 +143,25 @@ class ServerSideActionWdg(Button):
 		NetworkService.request(
 			None, url, secure=True,
 			successHandler=self.fetchSucceeded,
-			failureHandler=self.fetchFailed
+			failureHandler=NiceErrorAndThen(self.fetchFailed)
 		)
 
 	def fetchSucceeded(self, req):
 		if self.pendingFetches:
 			self.fetchNext()
 		else:
-			conf["mainWindow"].log("success", "Done")
+			conf["mainWindow"].log("success", self.actionData.get("success", "Done"))
 
 			self.removeClass("is-loading")
-			NetworkService.notifyChange(self.parent().parent().module)
 
-	def fetchFailed(self, req, code):
+			action = self.actionData.get("then", "reload-module")
+			if action == "reload-module":
+				NetworkService.notifyChange(self.parent().parent().module)
+			elif action == "reload-vi":
+				window.location.reload()
+
+	def fetchFailed(self):
 		self.pendingFetches = []
-		conf["mainWindow"].log("error", "Failed")
 		self.resetLoadingState()
 
 	def resetLoadingState(self):