+++ vi/widgets/list.py
@@ -63,7 +63,7 @@ class ListWidget(html5.Div):
 
 		myView = None
 		self.group = None
-		print(conf["modules"][module])
+
 		if conf["modules"] and module in conf["modules"].keys():
 			if filterID and "views" in conf["modules"][module] and conf["modules"][module]["views"]:
 				for v in conf["modules"][module]["views"]:
@@ -71,9 +71,10 @@ class ListWidget(html5.Div):
 						myView = v
 						break
 
-			if conf["modules"][module]["handler"] == "list.grouped":
-				if "group" in conf["modules"][module] and conf["modules"][module]["group"]:
-					self.group = kwargs["adminInfo"]["group"]
+			if conf["modules"][module]["handler"] == "list.grouped" \
+					or conf["modules"][module]["handler"].startswith("list.grouped."):
+				if "group" in kwargs and kwargs["group"]:
+					self.group = kwargs["group"]
 				else:
 					self.group = "all"
 
@@ -105,8 +106,8 @@ class ListWidget(html5.Div):
 
 		self.getAllActions(myView)
 
-		self.actionBar.setActions(self.getActions(), widget=self)
-		self.entryActionBar.setActions(self.getDefaultEntryActions(), widget=self)
+		self.actionBar.setActions(self.getActions(), widget=self,view=myView)
+		self.entryActionBar.setActions(self.getDefaultEntryActions(), widget=self,view=myView)
 
 		self.emptyNotificationDiv = html5.Div()
 		self.emptyNotificationDiv.prependChild(SvgIcon("icon-error-file", title="Currently no entries"))
@@ -162,6 +163,9 @@ class ListWidget(html5.Div):
 		:param kwargs: ListWidget Parameter
 		:return:
 		'''
+
+		if "indexes" in kwargs:
+			del kwargs["indexes"]  # ViUR 3.3 adminInfo contains this key
 
 		self.table = DataTable(checkboxes=self._checkboxes, indexes=self._indexes, *args, **kwargs)
 		self.widgetContent.appendChild(self.table)
@@ -230,7 +234,7 @@ class ListWidget(html5.Div):
 
 		cfg = None
 		if conf["modules"] and self.module in conf["modules"]:
-			cfg = conf["modules"][self.module]
+			cfg = conf["modules"][self.module].copy()
 
 		# update with view cfg
 		if view:
@@ -561,6 +565,10 @@ class ViewportListWidget(ListWidget):
 		Override explanation
 			- use ViewPort DataTable with rows parameter
 		'''
+
+		if "indexes" in kwargs:
+			del kwargs["indexes"]  # ViUR 3.3 adminInfo contains this key
+
 		self.table = ViewportDataTable(rows=self._batchSize,
 									   checkboxes=self._checkboxes,
 									   indexes=self._indexes,